hotfix/ 타입 변경 (#26)
* [수정]컬럼 수정

* [수정] 이름 누락 수정

* [수정] Double -> Long타입으로 변경

+++ src/main/java/com/AiFunding/ToBi/entity/AccountEntity.java
@@ -37,10 +37,10 @@ public class AccountEntity extends BaseCreateModifiedEntity implements Serializa
     private Long balance;
 
     @Column(name = "today_total_balance")
-    private Double todayTotalBalance;
+    private Long todayTotalBalance;
 
     @Column(name = "yesterday_total_balance")
-    private Double yesterdayTotalBalance;
+    private Long yesterdayTotalBalance;
 
     @Column(name = "ai_type")
     @NotNull
